# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: 多人网页气球飞行对战游戏 API
  description: |
    多人网页气球飞行对战游戏的后端 REST API 与 WebSocket 接口文档。
    认证方式为 HttpOnly Cookie（quickplay / session），管理接口使用 admin_token Cookie。
    错误响应遵循 RFC 7807 ProblemDetails 格式。
  version: "1.0.0"
  contact:
    name: Development Team
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://balloon-game.example.com
    description: 生产环境
security: []
...
paths:
...
  /api/v1/leaderboard:
    get:
      operationId: getLeaderboard
      summary: 公开排行榜
      description: scope=global|weekly，limit 默认 50。
      tags: [Stats]
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [global, weekly]
            default: global
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: 排行榜条目
          content:
            application/json:
              schema:
                type: object
                properties:
                  scope:
                    type: string
                  entries:
                    type: array
                    items:
                      type: object
        "400":
          description: 无效 scope
        "503":
          description: 服务不可用
  /api/v1/user/stats:
    get:
      operationId: getUserStats
      summary: 当前用户最佳分与对局数
      tags: [Stats]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 用户统计
          content:
            application/json:
              schema:
                type: object
                properties:
                  bestScore:
                    type: integer
                  gamesPlayed:
                    type: integer
        "401":
          description: 未认证

  # ─── User Data (GDPR) ─────────────────────────────────────────────
...
  /api/v1/admin/login:
    post:
      operationId: adminLogin
      summary: 管理员登录
      description: |
        使用管理员密码登录，成功后设置 admin_token Cookie（24h 有效期）。
        限速 10 次/5分钟。
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  format: password
                  description: 管理员密码
      responses:
        "200":
          description: 登录成功
          headers:
            Set-Cookie:
              description: "admin_token HttpOnly Cookie，24h 有效期"
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Admin JWT token
        "400":
          description: 请求体无效
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: 密码错误
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: 管理员未配置或密码未设置
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/admin/logout:
    post:
      operationId: adminLogout
      summary: 管理员登出
      description: |
        吊销当前 admin_token（按 jti 撤销），并清除 admin_token Cookie。
        需要管理员认证（任何已认证的管理员均可登出，无需 RBAC 校验）。
      tags: [Admin]
      security:
        - adminAuth: []
      responses:
        "200":
          description: 登出成功
          headers:
            Set-Cookie:
              description: "清除 admin_token Cookie（Max-Age=-1）"
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out"
        "401":
          description: 未认证（缺少或无效的 admin_token）
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/admin/config:
    get:
      operationId: adminGetConfig
      summary: 获取配置
      description: 获取全局配置。敏感字段（Resend API Key）已脱敏。需要管理员认证。
      tags: [Admin]
      security:
        - adminAuth: []
      responses:
        "200":
          description: 配置信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminConfigResponse"
        "401":
          description: 未认证（缺少或无效的 admin_token）
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: 配置不存在
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    patch:
      operationId: adminUpdateConfig
      summary: 更新配置
      description: |
        部分更新全局配置。仅提供的字段会被更新。
        Resend API Key 使用 AES-256-GCM 加密存储。
        管理员密码使用 bcrypt 哈希存储。
        需要管理员认证。
      tags: [Admin]
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminConfigUpdate"
      responses:
        "200":
          description: 配置更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Config updated"
        "400":
          description: 请求体无效
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: 配置不存在
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: 内部服务器错误（加密/哈希失败等）
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    put:
      operationId: adminUpdateConfigDeprecated
      summary: 更新配置（已弃用，请使用 PATCH）
      deprecated: true
      description: |
        已弃用：请使用 PATCH /api/v1/admin/config 代替。
        PUT 语义为全量替换，但实际行为为部分更新，不符合 RFC 规范。
        此端点将在 2027-01-01 后移除。
      tags: [Admin]
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminConfigUpdate"
      responses:
        "200":
          description: 配置更新成功
          headers:
            Deprecation:
              description: "标记此端点已弃用"
              schema:
                type: string
                example: "true"
            Sunset:
              description: "端点移除日期"
              schema:
                type: string
                example: "Sat, 01 Jan 2027 00:00:00 GMT"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Config updated"
        "400":
          description: 请求体无效
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: 配置不存在
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: 内部服务器错误（加密/哈希失败等）
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  # ─── Health ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: quickplay
      description: "用户 JWT access token Cookie（quickplay 或 session）"
    adminAuth:
      type: apiKey
      in: cookie
      name: admin_token
      description: "管理员 JWT Cookie，24h 有效期，role=admin"
  schemas:
...
    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: "错误类型标识 URI（RFC 7807）"
          example: "https://httpstatuses.com/400"
        title:
          type: string
          description: "人类可读的错误标题"
          example: "Bad Request"
        status:
          type: integer
          description: "HTTP 状态码"
          example: 400
        detail:
          type: string
          description: "具体错误详情"
          example: "Email is required"
        instance:
          type: string
          description: "出现问题的具体请求实例"

    # ─── Auth Schemas ────────────────────────────────────────────────
...
    AdminConfigResponse:
      type: object
      properties:
        emailEnabled:
          type: boolean
          description: "是否启用邮件功能"
          example: true
        resendApiKey:
          type: string
          description: "Resend API Key（已脱敏）"
          example: "••••••••"
        emailFrom:
          type: string
          description: "发件人邮箱地址"
          example: "noreply@example.com"
    AdminConfigUpdate:
      type: object
      description: "部分更新配置，仅提供的字段会被更新"
      properties:
        emailEnabled:
          type: boolean
          description: "是否启用邮件功能"
        resendApiKey:
          type: string
          description: "Resend API Key（明文，服务端自动 AES-256-GCM 加密存储；传 •••••••• 则不更新）"
        emailFrom:
          type: string
          description: "发件人邮箱地址"
        adminPassword:
          type: string
          format: password
          description: "新管理员密码（明文，服务端自动 bcrypt 哈希存储）"
        oldPassword:
          type: string
          format: password
          description: "当前管理员密码（修改 adminPassword 时必填，用于二次验证）"

    # ─── Health Schemas ──────────────────────────────────────────────
...